Move calendarUtils to bot src/util
