+++ matrix-meetings-bot/src/service/MeetingService.ts
@@ -61,9 +61,9 @@ import { powerLevelHelper } from '../model/PowerLevelHelper';
 import { RoomEventName } from '../model/RoomEventName';
 import { StateEventName } from '../model/StateEventName';
 import { WidgetType } from '../model/WidgetType';
-import { getForceDeletionTime } from '../shared';
 import { IMeetingChanges, meetingChangesHelper } from '../util/IMeetingChanges';
 import { templateHelper } from '../util/TemplateHelper';
+import { getForceDeletionTime } from '../util/calendarUtils';
 import { extractOxRrule } from '../util/extractOxRrule';
 import { migrateMeetingTime } from '../util/migrateMeetingTime';
 import { RoomMessageService } from './RoomMessageService';
